extract spawn app from health check test
